slack api: port from Slack Events API to Slack RTM API
Originally we thought RTM would be overkill for this bot.
However, using the Events API requires the bot to be publicly
accessible, and we need it to run internally.

The previous Events implementation used flask from the Slack
library. With RTM there is no longer a server to catch unhandled
Exceptions, so use a global try/catch in `handle_message`
to prevent any unhandled exception from stopping the bot.

There is no `app_mention` event in RTM. Added an env variable
for BOT_ID to reduce the RTM firehose to only messages we
care about, i.e. those @ the bot.