cluster-support-bot: Expose subscription managed and support values
Internal tooling does not support opening support cases for accounts
with no subscriptions.  Surface that information in the summary
display so folks looking at the cluster will know not to ask CEE to
reach out to the non-customer and request a must-gather or similar.

Also display managed state, because for clusters we manage on the
customer's behalf, we can ask for must-gather and similar internally,
and don't need CEE to pull the customer in.