telemetry: 'topk' -> 'group', and drop subscription range to 60h
'group' is more efficient, especially when we list the labels we're
looking for, because it can stop after it finds a single matching
entry instead of iterating through all the matching entries.  Although
if it needs to iterate through all the entries looking for possibly
different match groups, I dunno if that helps all that much.  Still,
unlikely to be less efficient.

Dropping the time range from a week to 60h should avoid errors we've
been seeing recently like:

  receive series from Addr: ...[name:"ruler_replica" value:"thanos-ruler-1" ] Mint: 1562803200000 Maxt: 1607630400000: rpc error: code = Aborted desc = fetch series for block 01ERQRVX3WQNWDFM24Z6SNF77E: expanded matching posting: get postings: read postings range: get range reader: The specified key does not exist.

I don't understand what's going on there, but time ranges of 63h and
larger seem to fail that way, while ranges of 62h and shorter succeed.
60 is a nice, round multiple of 12 that succeeds at the moment.