fix(account note type): per internal discussion, shifting note type from General to Technical
In order for the notes to get the appropriate attention (both preventing too much and too little),
the Technical Note type was created specifically with this bot in mind. The notes will once again
show up in cases (an important part of this process so support resources can see the notes from
engineering) but will Not trigger the Special Handling flag (A part of the Key Notes type that did
Not make sense for us to use). This change should not impact the usability of the bot in the
slightest.